store table and some typo in other tables

+++ store.php
@@ -19,16 +19,42 @@
         </div>
 
         <div class="bar">
+            <div id="mySidenav" class="sidenav">
+                <a href="javascript:void(0)" class="closebtn" onclick="closeNav()">&times;</a>
+                <a href= "actor.php">actor</a>
+                <a href= "address.php">address</a>
+                <a href= "category.php">category</a>
+                <a href= "city.php">city</a>
+                <a href= "country.php">country</a>
+                <a href= "customer.php">customer</a>
+                <a href= "film_actor.php">film_actor</a>
+                <a href= "film_category.php">film_category</a>
+                <a href= "film_text.php">film_text</a>
+                <a href= "film.php">film</a>
+                <a href= "inventory.php">inventory</a>
+                <a href= "language.php">language</a>
+                <a href= "payment.php">payment</a>
+                <a href= "rental.php">rental</a>
+                <a href= "staff.php">staff</a>
+                <a href= "store.php">store</a>
+                <a href= "payment&customer&staff.php">payment&customer&staff</a>
+                <a href= "film&film_text&film_category&category&language.php">film&film_text&film_category&category&language</a>
+                <a href= "country&city&address.php">country&city&address</a>
+            </div>
+
+            <div class="menu">
+                <span style="font-size:25px;cursor:pointer;color:white; text-align:left; display:inline;" onclick="openNav()">&#9776; MENU</span>
+            </div>
             <form action= "" method= "POST" style= 'display: inline;'>
-                <input type ="text" name= "search" placeholder="SEARCH BY ID" style= "border-radius: 5px;">
+                <input type ="text" name= "search" placeholder="SEARCH BY STORE ID" style= "border-radius: 5px;">
             </form>
             <form action= "" method= "POST" style= 'display: inline;'>
                 <input type = "submit" name= "reset" value= "RESET" class="button"> 
             </form>
-            <button id="insert" class= "button">INSERT</button>
-            <button id="update" class= "button">UPDATE</button>
-            <button id="delete" class= "button">DELETE</button>
-        </div>
+                <button id="insert" class= "button">INSERT</button>
+                <button id="update" class= "button">UPDATE</button>
+                <button id="delete" class= "button">DELETE</button>
+            </div>
 
         <?php
             echo "<table><thead><tr><th>Store_id</th><th>Address_id</th><th>Last_update</th></thead><tbody>"; 
@@ -81,7 +107,7 @@
                     <?php
                     if(isset($_POST['insert'])){
                         if(!empty($_POST['storeid'])&& !empty($_POST['addressid'])){
-                            if($_POST['storeid'] > 0){
+                            if((is_numeric($_POST['storeid']))&&($_POST['storeid'] > 0)){
                                 $storeid= $_POST['storeid'];
                                 $query= "SELECT store_id FROM store WHERE store_id = $storeid;";
                                 $result= mysqli_query($conn,$query);
@@ -96,16 +122,16 @@
                                         echo '<script> alert("DATA INSERTED SUCCESSFULLY!")</script>';
                                     }
                                     else
-                                        echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>'; 
+                                        echo '<script> alert("PREVIOUS INSERT FAILED! ADDRESS ID ENTERED DOES NOT EXIST, PLEASE CHECK THE ADDRESS TABLE FOR AN EXISTING ADDRESS ID")</script>'; 
 
                                     echo("<meta http-equiv='refresh' content='1'>");
                                 }
                                 else{
-                                    echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>';
+                                    echo '<script> alert("PREVIOUS INSERT FAILED! STORE ID ENTERED ALREADY EXISTS")</script>';
                                 }  
                             }
                             else{
-                                echo '<script> alert("PREVIOUS INSERT FAILED! INVALID ID ENTERED")</script>';
+                                echo '<script> alert("PREVIOUS INSERT FAILED! INVALID STORE ID ENTERED")</script>';
                             }
                         } 
                         else{
@@ -131,6 +157,7 @@
                     <?php
                     if(isset($_POST['update'])){
                         if(!empty($_POST['storeid'])&& !empty($_POST['addressid'])){
+                            if((is_numeric($_POST['storeid']))&&($_POST['storeid'] > 0)){
                             $storeid= $_POST['storeid'];
                             $query= "SELECT store_id FROM store WHERE store_id = $storeid;";
                             $result= mysqli_query($conn,$query);
@@ -145,14 +172,17 @@
                                 if($result)
                                     echo '<script> alert("DATA UPDATED SUCCESSFULLY!")</script>';
                                 else
-                                    echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>';
+                                    echo '<script> alert("PREVIOUS UPDATE FAILED! ADDRESS ID ENTERED DOES NOT EXIST, PLEASE CHECK THE ADDRESS TABLE FOR AN EXISTING ADDRESS ID")</script>';
                                 
                                 echo("<meta http-equiv='refresh' content='1'>");
                             }else{
-                                echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>';
+                                echo '<script> alert("PREVIOUS UPDATE FAILED! STORE ID ENTERED DOES NOT EXIST")</script>';
                             }
                                 
                         }else{
+                            echo '<script> alert("PREVIOUS UPDATE FAILED! INVALID STORE ID ENTERED")</script>';
+                        }
+                    }else{
                             echo '<script> alert("PREVIOUS UPDATE FAILED, PLEASE FILL ALL FIELDS!")</script>';
                         }
                     }
@@ -174,29 +204,34 @@
 
             <?php
                 if(isset($_POST['delete'])){
-                    if(!empty($_POST['storeid'])){
-                        $storeid= $_POST['storeid'];
-                        $query= "SELECT store_id FROM store WHERE store_id = $storeid;";
-                        $result= mysqli_query($conn,$query);
-                            
-                        if(mysqli_num_rows($result)==1){
+                    if ((is_numeric($_POST['storeid'])) && ($_POST['storeid'] > 0)){
+                        if(!empty($_POST['storeid'])){
                             $storeid= $_POST['storeid'];
-                            $delete = "DELETE FROM store WHERE store_id= '$storeid'; ";
-                            $result = mysqli_query($conn,$delete);
+                            $query= "SELECT store_id FROM store WHERE store_id = $storeid;";
+                            $result= mysqli_query($conn,$query);
                                 
-                            if($result){
-                                echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                            if(mysqli_num_rows($result)==1){
+                                $storeid= $_POST['storeid'];
+                                $delete = "DELETE FROM store WHERE store_id= '$storeid'; ";
+                                $result = mysqli_query($conn,$delete);
+                                    
+                                if($result){
+                                    echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                                }else{
+                                    echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                }
+                                    
+                                echo("<meta http-equiv='refresh' content='1'>");
                             }else{
-                                echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                echo '<script> alert("PREVIOUS DELETE FAILED! STORE ID ENTERED DOES NOT EXIST")</script>';
                             }
-                                
-                            echo("<meta http-equiv='refresh' content='1'>");
-                        }elseif (mysqli_num_rows($result)==0){
-                            echo '<script> alert("PREVIOUS DELETE FAILED! INVALID ID ENTERED")</script>';
                         }
-                                
-                    }else{
-                        echo '<script> alert("PREVIOUS DELETE FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                        else{
+                            echo '<script> alert("PREVIOUS DELETE FAILED! INVALID STORE ID ENTERED")</script>';
+                        }
+                    }
+                    else{
+                            echo '<script> alert("PREVIOUS DELETE FAILED, PLEASE FILL ALL FIELDS!")</script>';
                     }
                 }
             ?>
@@ -233,6 +268,13 @@
             function(){
                 document.querySelector('.delete').style.display= 'none';
             });
+            function openNav() {
+                document.getElementById("mySidenav").style.width = "620px";
+            }
+
+            function closeNav() {
+                document.getElementById("mySidenav").style.width = "0";
+            }
         </script>
     </body>
 </html>
